Initial release
Thois GUI was made because NEEO no longer has an APP in the Apple-store.
The NEEO-APP is still available in the Google PlayStore, but if your "an Apple-guy", there's no APP anymore  for you.
The only alternative for them is the native NEEO-GUI.

But, as far as I know, controlling of your devices is only possible for a small set of controlls: sliders and switches; buttons don't work.

Thats where this alterative GUI comes into play.  It allows the use of all controls(including buttons).
It does NOT contain any p[ossibility to change NEEO-settings as the native NEEO-GUI does a perfect job there (and recreating that is useless and very complex).
Use at your own risk, but  this solution has proven to be handy... for me.